Update the protocol.
